requirements.txt
joblib==1.3.2
matplotlib==3.8.3
pandas==1.4.4
python-dotenv==1.0.1
scikit_learn==1.4.0
pytest==8.0.0
pytest-cov==4.1.0
pickle4==0.0.1